Frontend/requirements.txt
ansi2html==1.8.0
cachelib==0.9.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
dash==2.11.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==2.2.5
Flask-Caching==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
nest-asyncio==1.5.6
numpy==1.25.0
packaging==23.1
pandas==2.0.2
plotly==5.15.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
retrying==1.3.4
six==1.16.0
tenacity==8.2.2
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
Werkzeug==2.2.3
